DEV: move tests not related to borsar.Clusters to test_clustering.py
